requirements.txt
flask==3.0.2
flask-cors==4.0.0
python-dotenv==1.0.1
tavily-python==0.3.1
aiohttp==3.9.3
asyncio==3.4.3
hypercorn==0.16.0
